ai-server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-dotenv==1.0.0
gymnasium==0.29.1
stable-baselines3==2.2.1
torch==2.1.1
numpy==1.24.3
pydantic==2.5.2
logging-config==1.0.3
asyncio-mqtt==0.13.0